2_0 branch merge

+++ lib/hash.c
@@ -335,10 +335,10 @@ hashPrime(int n)
     int I = sizeof(hash_primes) / sizeof(int);
     int i;
     int best_prime = hash_primes[0];
-    double min = fabs(log(n) - log(hash_primes[0]));
+    double min = fabs(log((double) n) - log((double) hash_primes[0]));
     double d;
     for (i = 0; i < I; i++) {
-	d = fabs(log(n) - log(hash_primes[i]));
+	d = fabs(log((double) n) - log((double) hash_primes[i]));
 	if (d > min)
 	    continue;
 	min = d;
